Split the cpu files into several.
