+++ CMakeLists.txt
@@ -44,6 +44,13 @@ set(CMAKE_CXX_STANDARD 11)
 
 set(SOURCE_FILES
 	src/cpu.cpp
+	src/cpu_instructions_misc_control.cpp
+	src/cpu_instructions_jumps_calls.cpp
+	src/cpu_instructions_8bits_load_store_move.cpp
+	src/cpu_instructions_16bits_load_store_move.cpp
+	src/cpu_instructions_8bits_arithmetic_logical.cpp
+	src/cpu_instructions_16bits_arithmetic_logical.cpp
+	src/cpu_instructions_8bits_rotation_shifts_bit.cpp
 	src/cpu.hpp
 	src/emulator.hpp
 	src/emulator.cpp
