+++ src/cpu.cpp
@@ -1,8 +1,4 @@
 #include "cpu.hpp"
-#include "instructions.hpp"
-#include "mmu.hpp"
-#include "utils.hpp"
-#include <cassert>
 #include <iostream>
 
 using namespace utils;
@@ -29,7 +25,7 @@ CPU::CPU(MMU& mmu_) : mmu(mmu_)
 
 int CPU::fetchDecodeAndExecute()
 {
-
+	// TODO: Review interrupts logic
 	if (ticksBeforeEnablingInterrupts > 0)
 	{
 		ticksBeforeEnablingInterrupts--;
@@ -85,14 +81,6 @@ void CPU::changeZeroValueFlag(byte value)
 
 CPU::~CPU() = default;
 
-void CPU::load16BitsValueInRegisters(byte& msbRegister, byte& lsbRegister)
-{
-	lsbRegister = mmu.read(pc);
-	msbRegister = mmu.read(pc + 1);
-	pc += 2;
-	lastInstructionTicks = 3;
-}
-
 void CPU::LD_XY_Z_N(byte& X, byte& Y, uint16_t Z)
 {
 	unsetFlag(CpuFlags::ZERO);
@@ -107,267 +95,10 @@ void CPU::LD_XY_Z_N(byte& X, byte& Y, uint16_t Z)
 	lastInstructionTicks = 3;
 }
 
-void CPU::load16BitsImmediateValueIntoRegister(uint16_t& reg)
-{
-	reg = mmu.readWord(pc);
-	pc += 2;
-	lastInstructionTicks = 3;
-}
-
 void CPU::LD_X_NNm(byte& X)
 {
 	X = mmu.read(mmu.readWord(pc));
 	pc += 2;
-	lastInstructionTicks = 4;
-}
-
-void CPU::load16BitsRegisterAtImmediateAddress(uint16_t reg)
-{
-	mmu.writeWord(mmu.readWord(pc), reg);
-	pc += 2;
-	lastInstructionTicks = 5;
-}
-
-void CPU::load8BitsRegisterAtImmediateAddress(byte reg)
-{
-	mmu.write(mmu.readWord(pc), reg);
-	pc += 2;
-	lastInstructionTicks = 4;
-}
-
-void CPU::loadImmediateValueInRegister(byte& reg)
-{
-	reg = mmu.read(pc);
-	pc++;
-	lastInstructionTicks = 2;
-}
-
-void CPU::loadValueToMemoryAtAddr(byte addrMsb, byte addrLsb, byte value)
-{
-	mmu.write(createAddrFromHighAndLowBytes(addrMsb, addrLsb), value);
-	lastInstructionTicks = 2;
-}
-
-void CPU::load8BitsImmediateValueAtMemoryAddress(byte addrMsb, byte addrLsb)
-{
-	byte value = mmu.read(pc);
-	pc++;
-	mmu.write(createAddrFromHighAndLowBytes(addrMsb, addrLsb), value);
-	lastInstructionTicks = 3;
-}
-
-void CPU::loadValueToMemoryAndIncrementAddr(byte& addrMsb, byte& addrLsb, byte value)
-{
-	mmu.write(createAddrFromHighAndLowBytes(addrMsb, addrLsb), value);
-	increment16BitsValueStoredIn8BitsValues(addrMsb, addrLsb);
-	lastInstructionTicks = 2;
-}
-
-void CPU::loadValueFromMemoryAndIncrementAddr(byte& reg, byte& addrMsb, byte& addrLsb)
-{
-	reg = mmu.read(createAddrFromHighAndLowBytes(addrMsb, addrLsb));
-	increment16BitsValueStoredIn8BitsValues(addrMsb, addrLsb);
-	lastInstructionTicks = 2;
-}
-
-void CPU::loadValueFromMemoryAndDecreaseAddr(byte& reg, byte& addrMsb, byte& addrLsb)
-{
-	reg = mmu.read(createAddrFromHighAndLowBytes(addrMsb, addrLsb));
-	decrement16BitsValueStoredIn8BitsValues(addrMsb, addrLsb);
-	lastInstructionTicks = 2;
-}
-
-void CPU::loadValueToMemoryAndDecreaseAddr(byte& addrMsb, byte& addrLsb, byte value)
-{
-	mmu.write(createAddrFromHighAndLowBytes(addrMsb, addrLsb), value);
-	decrement16BitsValueStoredIn8BitsValues(addrMsb, addrLsb);
-	lastInstructionTicks = 2;
-}
-
-void CPU::loadValueFromMemoryInto8BitsRegister(byte& reg, byte addrMsb, byte addrLsb)
-{
-	reg = mmu.read(createAddrFromHighAndLowBytes(addrMsb, addrLsb));
-	lastInstructionTicks = 2;
-}
-
-void CPU::load8BitsValueInRegister(byte& reg, byte value)
-{
-	reg = value;
-	lastInstructionTicks = 1;
-}
-
-void CPU::incrementRegistersValue(byte& msbRegister, byte& lsbRegister)
-{
-	lsbRegister++;
-	// If we overflowed the LSB then we increment the MSB
-	if (lsbRegister == 0)
-	{
-		msbRegister++;
-	}
-	lastInstructionTicks = 2;
-}
-
-void CPU::incrementRegisterValue(uint16_t& reg)
-{
-	reg++;
-	lastInstructionTicks = 2;
-}
-
-void CPU::incrementValueInMemoryAtAddr(byte addrMsb, byte addrLsb)
-{
-	unsetFlag(CpuFlags::SUBSTRACTION);
-	uint16_t addr = createAddrFromHighAndLowBytes(addrMsb, addrLsb);
-	byte oldvalue = mmu.read(addr);
-	byte value = oldvalue + 1;
-	mmu.write(addr, value);
-	changeZeroValueFlag(value);
-	setHalfCarryFlag(value == 0x10);
-
-	lastInstructionTicks = 3;
-}
-
-void CPU::decrementValueInMemoryAtAddr(byte addrMsb, byte addrLsb)
-{
-	setFlag(CpuFlags::SUBSTRACTION);
-	uint16_t addr = createAddrFromHighAndLowBytes(addrMsb, addrLsb);
-	byte oldvalue = mmu.read(addr);
-	byte value = oldvalue - 1;
-	mmu.write(addr, value);
-	changeZeroValueFlag(value);
-	setHalfCarryFlag(value == 0x08);
-	lastInstructionTicks = 3;
-}
-
-void CPU::incrementRegisterValue(byte& reg)
-{
-	unsetFlag(CpuFlags::SUBSTRACTION);
-
-	reg++;
-	setHalfCarryFlag(reg == 0x10);
-	changeZeroValueFlag(reg);
-	lastInstructionTicks = 1;
-}
-
-void CPU::decrementRegisterValue(byte& reg)
-{
-	setFlag(CpuFlags::SUBSTRACTION);
-
-	reg--;
-	setHalfCarryFlag(reg == 0x08);
-	changeZeroValueFlag(reg);
-	lastInstructionTicks = 1;
-}
-
-void CPU::decrementRegistersValue(byte& msgRegister, byte& lsbRegister)
-{
-	uint16_t xy = (msgRegister << 8) | lsbRegister;
-	xy--;
-	msgRegister = (xy >> 8);
-	lsbRegister = (xy & 0x00FF);
-	lastInstructionTicks = 2;
-}
-
-void CPU::decrementRegisterValue(uint16_t& reg)
-{
-	reg--;
-	lastInstructionTicks = 2;
-}
-
-void CPU::rotateRegisterLeftCircular(byte& reg)
-{
-	unsetFlag(CpuFlags::SUBSTRACTION);
-	unsetFlag(CpuFlags::ZERO);
-	unsetFlag(CpuFlags::HALF_CARRY);
-
-	/* Set the carry flag to the highest bit of X */
-	setCarryFlag((reg & 0x80) > 0);
-	/* Rotate the value circularly */
-	reg = (reg << 1) | (reg >> 7);
-	// TODO: Confirm that the Zero flag should also be set in the standard instruction set
-	changeZeroValueFlag(reg);
-	lastInstructionTicks = 1;
-}
-
-void CPU::rotateValueInMemoryLeftCircular(byte addrMsb, byte addrLsb)
-{
-	uint16_t addr = createAddrFromHighAndLowBytes(addrMsb, addrLsb);
-	byte value = mmu.read(addr);
-	rotateRegisterLeftCircular(value);
-	mmu.write(addr, value);
-	lastInstructionTicks = 4;
-}
-
-void CPU::rotateRegisterRightCircular(byte& reg)
-{
-	unsetFlag(CpuFlags::SUBSTRACTION);
-	unsetFlag(CpuFlags::ZERO);
-	unsetFlag(CpuFlags::HALF_CARRY);
-
-	/* Set the carry flag to the lowest bit of X */
-	setCarryFlag((reg & 0x01) > 0);
-	/* Rotate the value circularly */
-	reg = (reg << 7) | (reg >> 1);
-	// TODO: Confirm that the Zero flag should also be set in the standard instruction set
-	changeZeroValueFlag(reg);
-	lastInstructionTicks = 1;
-}
-
-void CPU::rotateValueInMemoryRightCircular(byte addrMsb, byte addrLsb)
-{
-	uint16_t addr = createAddrFromHighAndLowBytes(addrMsb, addrLsb);
-	byte value = mmu.read(addr);
-	rotateRegisterRightCircular(value);
-	mmu.write(addr, value);
-	lastInstructionTicks = 4;
-}
-
-void CPU::rotateRegisterLeftExtended(byte& reg)
-{
-	unsetFlag(CpuFlags::SUBSTRACTION);
-	unsetFlag(CpuFlags::ZERO);
-	unsetFlag(CpuFlags::HALF_CARRY);
-
-	byte currentCarry = isFlagSet(CpuFlags::CARRY);
-	/* Set the carry flag to the highest bit of reg */
-	setCarryFlag((reg & 0x80) > 0);
-	/* Rotate the accumulator and set the last bit to the original carry flag */
-	reg = (reg << 1) | currentCarry;
-	// TODO: Confirm that the Zero flag should also be set in the standard instruction set
-	changeZeroValueFlag(reg);
-	lastInstructionTicks = 2;
-}
-
-void CPU::rotateValueInMemoryLeft(byte addrMsb, byte addrLsb)
-{
-	uint16_t addr = createAddrFromHighAndLowBytes(addrMsb, addrLsb);
-	byte value = mmu.read(addr);
-	rotateRegisterLeftExtended(value);
-	mmu.write(addr, value);
-	lastInstructionTicks = 4;
-}
-
-void CPU::rotateRegisterRightExtended(byte& reg)
-{
-	unsetFlag(CpuFlags::SUBSTRACTION);
-	unsetFlag(CpuFlags::ZERO);
-	unsetFlag(CpuFlags::HALF_CARRY);
-
-	byte currentCarry = isFlagSet(CpuFlags::CARRY);
-	/* Set the carry flag to the lowest bit of X */
-	setCarryFlag((reg & 0x01) > 0);
-	/* Rotate the accumulator and set the highest bit to the original carry flag */
-	reg = (currentCarry << 7) | (reg >> 1);
-	// TODO: Confirm that the Zero flag should also be set in the standard instruction set
-	changeZeroValueFlag(reg);
-	lastInstructionTicks = 2;
-}
-
-void CPU::rotateValueInMemoryRight(byte addrMsb, byte addrLsb)
-{
-	uint16_t addr = createAddrFromHighAndLowBytes(addrMsb, addrLsb);
-	byte value = mmu.read(addr);
-	rotateRegisterRightExtended(value);
-	mmu.write(addr, value);
 	lastInstructionTicks = 4;
 }
 
@@ -495,31 +226,6 @@ void CPU::SET_X_YZm(byte X, byte Y, byte Z)
 	lastInstructionTicks = 4;
 }
 
-void CPU::addTwo8BitsRegistersToTwo8BitsRegisters(byte& resultRegMsb, byte& resultRegLsb, byte valueRegMsb,
-                                                  byte valueRegLsb)
-{
-	unsetFlag(CpuFlags::SUBSTRACTION);
-	uint16_t origValue = (resultRegMsb << 8) | resultRegLsb;
-	uint32_t result = origValue;
-	result += (valueRegMsb << 8) | valueRegLsb;
-	setCarryFlag(result > 0xFFFF);
-	setHalfCarryFlag(origValue <= 0xFFF && result > 0xFFF);
-	resultRegMsb = ((result & 0xFF00) >> 8);
-	resultRegLsb = (result & 0xFF);
-	lastInstructionTicks = 2;
-}
-
-void CPU::add8BitsValueTo8BitsRegister(byte& reg, byte value)
-{
-	unsetFlag(CpuFlags::SUBSTRACTION);
-	uint16_t result = reg + value;
-	setCarryFlag(result > 0xFF);
-	setHalfCarryFlag(reg <= 0x0F && result > 0x0F);
-	reg = static_cast<byte>(result);
-	changeZeroValueFlag(reg);
-	lastInstructionTicks = 1;
-}
-
 void CPU::ADD_SP_X(sbyte X)
 {
 	unsetFlag(CpuFlags::SUBSTRACTION);
@@ -529,249 +235,6 @@ void CPU::ADD_SP_X(sbyte X)
 	setCarryFlag(value > sp);
 	setHalfCarryFlag((value <= 0x00FF) && (sp > 0x00FF));
 	lastInstructionTicks = 4;
-}
-
-void CPU::addImmediateValueTo8BitsRegister(byte& reg)
-{
-	add8BitsValueTo8BitsRegister(reg, mmu.read(pc++));
-	lastInstructionTicks = 2;
-}
-
-void CPU::add8BitsValueAndCarryTo8BitsRegister(byte& reg, byte value)
-{
-	unsetFlag(CpuFlags::SUBSTRACTION);
-	uint16_t result = reg + value;
-	if (isFlagSet(CpuFlags::CARRY))
-	{
-		result++;
-	}
-	setCarryFlag(result > 0xFF);
-	setHalfCarryFlag(reg <= 0x0F && result > 0x0F);
-	reg = static_cast<byte>(result);
-	changeZeroValueFlag(reg);
-	lastInstructionTicks = 1;
-}
-
-void CPU::addImmediateValueAndCarryTo8BitsRegister(byte& reg)
-{
-	add8BitsValueAndCarryTo8BitsRegister(reg, mmu.read(pc++));
-	lastInstructionTicks = 2;
-}
-
-void CPU::addValueFromMemoryTo8BitsRegister(byte& reg, byte addrMsb, byte addrLsb)
-{
-	add8BitsValueTo8BitsRegister(reg, mmu.read(createAddrFromHighAndLowBytes(addrMsb, addrLsb)));
-	lastInstructionTicks = 2;
-}
-
-void CPU::addValueFromMemoryAndCarryTo8BitsRegister(byte& reg, byte addrMsb, byte addrLsb)
-{
-	add8BitsValueAndCarryTo8BitsRegister(reg, mmu.read(createAddrFromHighAndLowBytes(addrMsb, addrLsb)));
-	lastInstructionTicks = 2;
-}
-
-void CPU::substract8BitsValueFrom8BitsRegister(byte& reg, byte value)
-{
-	setFlag(CpuFlags::SUBSTRACTION);
-	uint16_t result = reg - value;
-	setCarryFlag(result > 0xFF);
-	setHalfCarryFlag(reg > 0x0F && result <= 0x0F);
-	reg = static_cast<byte>(result);
-	changeZeroValueFlag(reg);
-	lastInstructionTicks = 1;
-}
-
-void CPU::substractImmediateValueFrom8BitsRegister(byte& reg)
-{
-	substract8BitsValueFrom8BitsRegister(reg, mmu.read(pc++));
-	lastInstructionTicks = 2;
-}
-
-void CPU::substractValueInMemoryFrom8BitsRegister(byte& reg, byte addrMsb, byte addrLsb)
-{
-	substract8BitsValueFrom8BitsRegister(reg, mmu.read(createAddrFromHighAndLowBytes(addrMsb, addrLsb)));
-	lastInstructionTicks = 2;
-}
-
-void CPU::sub8BitsValueAndCarryTo8BitsRegister(byte& reg, byte value)
-{
-	setFlag(CpuFlags::SUBSTRACTION);
-	uint16_t result = reg - value;
-	if (isFlagSet(CpuFlags::CARRY))
-	{
-		result--;
-	}
-	setCarryFlag(result > 0xFF);
-	setHalfCarryFlag(reg > 0x0F && result <= 0x0F);
-	reg = static_cast<byte>(result);
-	changeZeroValueFlag(reg);
-	lastInstructionTicks = 1;
-}
-
-void CPU::subImmediateValueAndCarryTo8BitsRegister(byte& reg)
-{
-	sub8BitsValueAndCarryTo8BitsRegister(reg, mmu.read(pc++));
-	lastInstructionTicks = 2;
-}
-
-void CPU::subValueFromMemoryAndCarryTo8BitsRegister(byte& reg, byte addrMsb, byte addrLsb)
-{
-	sub8BitsValueAndCarryTo8BitsRegister(reg, mmu.read(createAddrFromHighAndLowBytes(addrMsb, addrLsb)));
-	lastInstructionTicks = 2;
-}
-
-void CPU::logicalAndBetweenAccumulatorAnd8BitsRegister(byte value)
-{
-	unsetFlag(CpuFlags::SUBSTRACTION);
-	setFlag(CpuFlags::HALF_CARRY);
-	unsetFlag(CpuFlags::CARRY);
-	a &= value;
-	changeZeroValueFlag(a);
-	lastInstructionTicks = 1;
-}
-
-void CPU::logicalAndBetweenAccumulatorAndValueInMemory(byte addrMsb, byte addrLsb)
-{
-	logicalAndBetweenAccumulatorAnd8BitsRegister(mmu.read(createAddrFromHighAndLowBytes(addrMsb, addrLsb)));
-	lastInstructionTicks = 2;
-}
-
-void CPU::logicalAndBetweenAccumulatorAndImmediateValue()
-{
-	logicalAndBetweenAccumulatorAnd8BitsRegister(mmu.read(pc++));
-	lastInstructionTicks = 2;
-}
-
-void CPU::logicalXorBetweenAccumulatorAnd8BitsRegister(byte value)
-{
-	unsetFlag(CpuFlags::SUBSTRACTION);
-	unsetFlag(CpuFlags::HALF_CARRY);
-	unsetFlag(CpuFlags::CARRY);
-	a ^= value;
-	changeZeroValueFlag(a);
-	lastInstructionTicks = 1;
-}
-
-void CPU::logicalXorBetweenAccumulatorAndValueInMemory(byte addrMsb, byte addrLsb)
-{
-	logicalXorBetweenAccumulatorAnd8BitsRegister(mmu.read(createAddrFromHighAndLowBytes(addrMsb, addrLsb)));
-	lastInstructionTicks = 2;
-}
-
-void CPU::logicalXorBetweenAccumulatorAndImmediateValue()
-{
-	logicalXorBetweenAccumulatorAnd8BitsRegister(mmu.read(pc++));
-	lastInstructionTicks = 2;
-}
-
-void CPU::logicalOrBetweenAccumulatorAnd8BitsRegister(byte value)
-{
-	unsetFlag(CpuFlags::SUBSTRACTION);
-	unsetFlag(CpuFlags::HALF_CARRY);
-	unsetFlag(CpuFlags::CARRY);
-	a |= value;
-	changeZeroValueFlag(a);
-	lastInstructionTicks = 1;
-}
-
-void CPU::logicalOrBetweenAccumulatorAndValueInMemory(byte addrMsb, byte addrLsb)
-{
-	logicalOrBetweenAccumulatorAnd8BitsRegister(mmu.read(createAddrFromHighAndLowBytes(addrMsb, addrLsb)));
-	lastInstructionTicks = 2;
-}
-
-void CPU::logicalOrBetweenAccumulatorAndImmediateValue()
-{
-	logicalOrBetweenAccumulatorAnd8BitsRegister(mmu.read(pc++));
-	lastInstructionTicks = 2;
-}
-
-void CPU::compareAccumulatorAndRegister(byte value)
-{
-	setFlag(CpuFlags::SUBSTRACTION);
-	setFlagIfTrue(a == value, CpuFlags::ZERO);
-	setHalfCarryFlag(a > value);
-	setCarryFlag(a < value);
-	lastInstructionTicks = 1;
-}
-
-void CPU::compareAccumulatorAndValueInMemory(byte addrMsb, byte addrLsb)
-{
-	compareAccumulatorAndRegister(mmu.read(createAddrFromHighAndLowBytes(addrMsb, addrLsb)));
-	lastInstructionTicks = 2;
-}
-
-void CPU::compareAccumulatorAndImmediateValue()
-{
-	compareAccumulatorAndRegister(mmu.read(pc++));
-	lastInstructionTicks = 2;
-}
-
-void CPU::popMemoryIntoRegisterPair(byte& regMsb, byte& regLsb)
-{
-    regLsb = mmu.read(sp);
-    regMsb = mmu.read(sp + 1);
-	sp += 2;
-	lastInstructionTicks = 3;
-}
-
-void CPU::returnInstruction()
-{
-	// We unpop the address of the caller
-	// From the stack and set the pc to the value
-	uint16_t addr = mmu.readWord(sp);
-	sp += 2;
-	pc = addr;
-	lastInstructionTicks = 4;
-}
-
-void CPU::returnInstructionConditional(bool cond)
-{
-	if (cond)
-	{
-		returnInstruction();
-		lastInstructionTicks = 5;
-	}
-	else
-	{
-		lastInstructionTicks = 2;
-	}
-}
-
-void CPU::returnInstructionAfterInterrupt()
-{
-	returnInstruction();
-	// TODO:
-	// Revisit what to do after implementing interrups
-	interrupts = false;
-}
-
-void CPU::jumpConditional(bool condition)
-{
-	uint16_t addr = mmu.readWord(pc);
-	pc += 2;
-	if (condition)
-	{
-		pc = addr;
-		lastInstructionTicks = 4;
-	}
-	else
-	{
-		lastInstructionTicks = 3;
-	}
-}
-
-void CPU::jump()
-{
-	uint16_t addr = mmu.readWord(pc);
-	pc = addr;
-	lastInstructionTicks = 4;
-}
-
-void CPU::jumpToAddrIn16BitsRegister(byte addrMsb, byte addrLsb)
-{
-	pc = createAddrFromHighAndLowBytes(addrMsb, addrLsb);
-	lastInstructionTicks = 1;
 }
 
 void CPU::CALL_X_NN(bool cond)
@@ -829,78 +292,6 @@ void CPU::LD_X_Ym(byte& X, byte Y)
 {
 	X = mmu.read(0xFF00 | Y);
 	lastInstructionTicks = 2;
-}
-
-void CPU::jumpRelativeConditional(bool condition)
-{
-	auto offset = static_cast<int8_t>(mmu.read(pc));
-	pc++;
-	if (condition)
-	{
-		pc += offset;
-		lastInstructionTicks = 3;
-	}
-	else
-	{
-		lastInstructionTicks = 2;
-	}
-}
-
-void CPU::jumpRelative()
-{
-	auto offset = static_cast<int8_t>(mmu.read(pc));
-	pc++;
-	pc += offset;
-	lastInstructionTicks = 3;
-}
-
-void CPU::decimalAdjustAccumulator()
-{
-	byte valueToAdd = 0;
-	if (((a & 0x0F) > 9) || isFlagSet(CpuFlags::HALF_CARRY))
-	{
-		valueToAdd += 0x06;
-	}
-
-	if (((a >> 4) > 9) || isFlagSet(CpuFlags::CARRY))
-	{
-		valueToAdd += 0x60;
-		setFlag(CpuFlags::CARRY);
-	}
-	else
-	{
-		unsetFlag(CpuFlags::CARRY);
-	}
-
-	a += valueToAdd;
-
-	setFlagIfTrue(a == 0, CpuFlags::ZERO);
-	unsetFlag(CpuFlags::HALF_CARRY);
-	lastInstructionTicks = 1;
-}
-
-void CPU::complementRegister(byte& reg)
-{
-	reg = ~reg;
-	setFlag(CpuFlags::SUBSTRACTION);
-	setFlag(CpuFlags::HALF_CARRY);
-	lastInstructionTicks = 1;
-}
-
-void CPU::setCarryFlagInstruction()
-{
-	unsetFlag(CpuFlags::HALF_CARRY);
-	unsetFlag(CpuFlags::SUBSTRACTION);
-	setFlag(CpuFlags::CARRY);
-	lastInstructionTicks = 1;
-}
-
-void CPU::invertsCarryFlag()
-{
-	unsetFlag(CpuFlags::HALF_CARRY);
-	unsetFlag(CpuFlags::SUBSTRACTION);
-	setCarryFlag(!isFlagSet(CpuFlags::CARRY));
-	lastInstructionTicks = 1;
 }
 
 /*
@@ -2982,26 +2373,9 @@ void CPU::setCarryFlag(bool state)
 	setFlagIfTrue(state, CpuFlags::CARRY);
 }
 
-void CPU::NoOperation()
-{
-	lastInstructionTicks = 1;
-}
-
 byte CPU::getFlag() const
 {
 	return f;
-}
-
-void CPU::rotateRegisterLeftCircularExtended(byte& reg)
-{
-	rotateRegisterLeftCircular(reg);
-	lastInstructionTicks = 2;
-}
-
-void CPU::rotateRegisterRightCircularExtended(byte& reg)
-{
-	rotateRegisterRightCircular(reg);
-	lastInstructionTicks = 2;
 }
 
 void CPU::resetFlags()
