+++ src/cpu_instructions_misc_control.cpp
@@ -1,0 +1,6 @@
+#include "cpu.hpp"
+
+void CPU::NoOperation()
+{
+    lastInstructionTicks = 1;
+}
